events: move pie tests to hover_pie_test.js
